pyproject.toml
[tool.poetry]
name = "ivcap_sdk_service"
version = "0.2.0"
description = "SDK library for building services for the IVCAP platform"
authors = ["Max Ott <max.ott@csiro.au>, Tim Erwin <tim.erwin@csiro.au>"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
importlib-metadata = "^5.0.0"
PyYAML = "^6.0"
dataclass-wizard = "^0.22.1"
validators = "^0.20.0"
requests = "^2.28"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
Sphinx = "^5.2.3"
myst-nb = "^0.17.1"
autoapi = "^2.0.1"
sphinx-autoapi = "^2.0.0"
sphinx-rtd-theme = "^1.0.0"
licenseheaders = "^0.8.8"

[tool.poetry.scripts]
ivcap-service = 'ivcap_service:run'

[tool.semantic_release]
version_variable = "pyproject.toml:version" # version location
branch = "main"                             # branch to make releases of
changelog_file = "CHANGELOG.md"             # changelog file
build_command = "poetry build"              # build dists
dist_path = "dist/"                         # where to put dists
upload_to_release = true                    # auto-create GitHub release
upload_to_pypi = false                      # don't auto-upload to PyPI
remove_dist = false                         # don't remove dists
patch_without_tag = true                    # patch release by default

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
